continue json stream parsing

+++ plugins/org.chromium.sdk/src/org/chromium/sdk/internal/shellprotocol/tools/protocol/input/ToolsMessage.java
@@ -16,7 +16,7 @@ import java.util.List;
 public interface ToolsMessage {
   String command();
 
-  long result();
+  int result();
 
   @JsonOptionalField
   Data data();
